Refactor code in page-load.js module
