+++ src/index.js
@@ -2,17 +2,16 @@ import "../assets/css/reset.css";
 import "../assets/css/navbar.css";
 import "../assets/css/style.css";
 
+import { pageLoad } from "./page-load";
 
-import { home, menu, contact } from "./home";
-
-home();
+pageLoad.home();
 
 const homeLink = document.getElementById('home');
 const menuLink = document.getElementById('menu');
 const contactLink = document.getElementById('contact');
 
-menuLink.addEventListener('click', menu);
+menuLink.addEventListener('click', pageLoad.menu);
 
-homeLink.addEventListener('click', home);
+homeLink.addEventListener('click', pageLoad.home);
 
-contactLink.addEventListener('click', contact);
+contactLink.addEventListener('click', pageLoad.contact);
